Re-route Organizer's files

+++ organizer/php-scripts/event-remover.php
@@ -4,23 +4,35 @@ session_start();
 require_once "../../services/connectDB.php";
 require_once "Event.php";
 
+// Send them back to index if they're not organizer
+if($_SESSION["ROLE"] !== "OR"){
+    header("Location: index.php");
+}
+
 if (!empty($_GET)) {
     $eventID = $_GET["eid"];
 
-    // delete thumbnail
-    $query = "DELETE FROM `event_image` WHERE `eventID` = $eventID";
+    // get orgID from eventID, to prevent some brute force breach.
+    $query = "SELECT `orgID` FROM `event` WHERE `eventID` = $eventID";
     $statement = $conn->prepare($query);
-    $statement->execute();
+    $result = $statement->fetch(PDO::FETCH_ASSOC);
+    $orgIDFromEvent = $result['orgID'];
+    if($orgIDFromEvent == $_SESSION['ID']){
+        // delete thumbnail
+        $query = "DELETE FROM `event_image` WHERE `eventID` = $eventID";
+        $statement = $conn->prepare($query);
+        $statement->execute();
 
-    // delete survey link
-    $query = "DELETE FROM `event_survey_link` WHERE `eventID` = $eventID";
-    $statement = $conn->prepare($query);
-    $statement->execute();
+        // delete survey link
+        $query = "DELETE FROM `event_survey_link` WHERE `eventID` = $eventID";
+        $statement = $conn->prepare($query);
+        $statement->execute();
 
-    // delete event
-    $query = "DELETE FROM `event` WHERE `eventID` = $eventID";
-    $statement = $conn->prepare($query);
-    $statement->execute();
+        // delete event
+        $query = "DELETE FROM `event` WHERE `eventID` = $eventID";
+        $statement = $conn->prepare($query);
+        $statement->execute();
+    }
 }
